Learnt about State Management in Jetpack Compose
