+++ app/build.gradle.kts
@@ -73,4 +73,9 @@ dependencies {
     // Window Size Class
     implementation("com.google.android.material:material:1.12.0")
 
-    implementation("androidx.compose.material3:material3-window-size-class:1.2.1")}
+    implementation("androidx.compose.material3:material3-window-size-class:1.2.1")
+
+    // ViewModel
+    implementation("androidx.lifecycle:lifecycle-viewmodel-compose:2.8.5")
+
+}
