+++ app/src/main/java/com/rajit/learning_compose/sootheapp/model/DrawableStringPair.kt
@@ -1,0 +1,9 @@
+package com.rajit.learning_compose.sootheapp.model
+
+import androidx.annotation.DrawableRes
+import androidx.annotation.StringRes
+
+data class DrawableStringPair(
+    @DrawableRes val drawable: Int,
+    @StringRes val text: Int
+)
