+++ app/src/main/java/com/rajit/learning_compose/wellnessapp/WellnessViewModel.kt
@@ -1,0 +1,28 @@
+package com.rajit.learning_compose.wellnessapp
+
+import androidx.compose.runtime.toMutableStateList
+import androidx.lifecycle.ViewModel
+
+class WellnessViewModel: ViewModel() {
+
+    private val _tasks = getWellnessTasks().toMutableStateList()
+    val tasks: List<WellnessTask> get() = _tasks
+
+    fun remove(item: WellnessTask) {
+        _tasks.remove(item)
+    }
+
+    fun changeTaskChecked(item: WellnessTask, checked: Boolean) {
+        _tasks.find { it.id == item.id }?.let { task ->
+            task.checked = checked
+        }
+    }
+
+    private fun getWellnessTasks() = List(30) { i ->
+        WellnessTask(
+            id = i,
+            label = "Task #$i"
+        )
+    }
+
+}
